Manual distillation test via CLI (#176)
This is mainly copied from Middleman with very little modification. I plan to have a follow-up which will do the following:

* Use a proper and richer logging
* Avoid custom browser code and reuse the existing abstraction
* Switch to idiomatic Pydantic

For now, try this by running:

```bash
./getgather.py run goodreads.com/review
```

and you'll be prompted for email and password. If nothing goes wrong,
the list of books will be printed out in its original HTML table.

And to run a simple distillation:

```bash
./getgather.py distill https://www.goodreads.com/user/sign_in
```