Distillation-based MCP tool to get NYTimes Bestsellers (#211)
This is just the first step. The next step is to incorporate some
parsing so that the raw-ish HTML can be turned into JSON.

To try, launch as usual with `npm run dev` and use MCP Inspector to
invoke the tool named `nytimes_get_bestsellers_list`. Alternatively, use
Claude or other MCP client.