Distilled page, to prepare for sign-in flow processing (#251)
This is not yet fully completed nor connected to the MCP tool. It is
merely the port of the Middleman's implementation of the concept.

To verify the initial functionality, launch with `npm run dev` and then
open the following `localhost:5173/dpage?location=bbc.com/saved`.

This will trigger the distillation loop and that dpage ("distilled
page") will start with the BBC sign in (email, and then password), and
the finally end up with message to close the window.

While this happens, expect Patchright/Chromium to launch and to navigate
the BBC website, according to the progression of the distillation.