Adopt the input value when it is being filled
This is necessary to avoid the potential infinite loop, e.g. in the case
of repeated incorrect credentials. It is a port of the same mechanism in
Middleman (https://github.com/mcp-getgather/middleman/pull/30).

The idea is that when an input field is filled, either on the CLI or
from the distilled page, the value for that field also changes the
distilled version of the page. This way, when that page is submitted
(via form) to the upstream server and the server comes back with another
sign-in form, it will not be considered the same thing (they are not).

To verify, run the usual CLI: `./getgather.py run goodreads.com/review`.

Before this change, the distillation loop never considers that the
second incorrect passsword alert as a different once.

After this change, everything works as expected. Everytime Goodreads
detect the incorrect password, the distilled version continues to be
delivered to the user.

To try it on the MCP server itself, use BBC since it's the one fully
migrated to distillation already.