Have a proper default scheme when creating a distilled page
This is a follow-up to PR #385. Apparently the default scheme can't be
always `https`, particularly when the Host is local.

To verify, `npm run dev`, connect MCP inspector, and invoke the tool
`bbc_get_saved_article`. The link returned should should the base URL
`http://localhost:23456`.

Before this PR, the base URL will use `https`, which leads to a failure.