Fix autoimport by checking for the MCP class in the module source (#398)
Eventually when everything is migrated to distillation, we may not need
this. But for now, avoid blind module import by checking first
heuristically if the module is a legimate MCP class.

To verify, run the MCP server, connect MCP inspector, list some tools.
All the available tools should be still enumerated correctly.