Allow sharing of VNC sessions to enable multiple VNC connections (#420)
Before this PR, if one person is opening the `/live` desktop view,
nobody else can connect to that, since by default the VNC server limits
to a single connection.

After this PR, by enabling the shared connection, multiple persons can
go to the `/live` view and interact with the desktop all at the same
time.