Take into account the hostname when creating a distilled page
This is necessary so that the MCP Gateway can route the tool call to the
right instance of MCP server, similar to the way it's done for hosted
links.

To verify, run by setting the hostname, e.g.:

```
HOSTNAME=foobar npm run dev
```

and then connect MCP Inspector, invoke a tool that requires
distillation, e.g. `bbc_get_saved_articles`. The result is the request
to continue with the sign in at a special distilled page, which has
`foobar` as the prefix.