Prioritize checking for termination in the distillation loop (#446)
This is to prepare for the separation of autoclick and form submission.

See https://github.com/mcp-getgather/middleman/pull/40.

To verify, run the distillation manually on the CLI, e.g.:

```bash
./getgather.py run goodreads.com/review
```

Alternatively, launch the MCP server as usual, connect MCP Inspector,
try an MCP tool such as `bbc_get_saved_articles`, and follow the sign-in
steps as instructed.