Properly create a new page from the browser context
This is to account for resource filtering, see PR #503 and #516.

To account for resources filtering (from the blocklists), a new page
needs to be immediately constructed so that the filtering takes effect.
This is also an opportunity to use the old page for a sanity check by
going to ifconfig.me (to reveal the public IP address, among others).

This applies to both the sign-in orchestrator and distillation.

To verify, launch as usual to /mcp-shopping and /mcp-media.
For the former, try the Amazon tool to search for a product or to get
the order history (sign-in orchestrator) and for the latter, use the BBC
tool (distillation).

Before this PR, Chromium opened with that Amazon/BBC pages will be
filled its images and other heavy resources, indicating that resources
filtering did not apply to that page.

After this PR, that initial page of Amazon/BBC has no images at all (as
well as custom fonts, etc, including trackers and ads). This is the
ultimate proof that resources are filtered properly.