Block certain URLs based on the preloaded blocklists (#516)
Any `blocklists-*.txt` will be treated as a list of domains to be
blocked (one domain per line).

To verify, run the MCP server with LOG_LEVEL=DEBUG, connect to /mcp-books
with the MCP Inspector, and test an MCP tool like Goodreads. Resources
with URLs in the blocklists should be marked as DENY.