Close the page after distillation is completed
Now that the page is always a fresh new one, the tool needs to close it
explicitly after the distillation process.

To test, run the MCP server and connect MCP Inspector. Check with BBC
tool (which is 100% distillation-based) and also Goodreads (sign-in with
orchestraton, then handed off to distillation-based extraction). In both
cases, the page which has the information (saved article for BBC,
bookshelf list for Goodreads, etc) should not stay in the browser window
forever, as it gets closed right after extraction.