Remove obsolete debugging in dpage (#568)
Originally you can hit `localhost:23456/dpage?location=...` to start the
distillation on that URL (location). However, the proper way to debug is
to use the MCP inspector which invokes the right MCP tool. And since
it's so easy anway to use MCP inspector, there's almost no meaningful
benefit anymore using the above `dpage?location=` trick.

This PR removes it, and we should use MCP inspector for proper
debugging.