Use a neutral default title in distilled page (#590)
In case the distillation pattern does not supply the HTML title, use the
neutral "Sign In" instead of "GetGather". This distilled page might be
used by some demo apps (e.g. Page Turner) and it can be confusing for
the user to see the text "GetGather", which has nothing to do with the
app itself.

Beside, the distilled page is solely used for the sign-in process
anyway, hence the default title "Sign In" is sensible enough.