Reduce distillation noise by not printing the distillation result
We can't use logger since it may contain sensitive information. And in
many cases it's a large content of HTML, printing it out stdout creates
a lot of unnecessary noise.

When debugging some distillation patterns, it's easy enough to do that
with Middleman and therefore we don't have to duplicate the
functionality here.