Initial Zendriver-based MCP tool for NPR headlines (#608)
This is an experiment to gradually migrate to Zendriver. The NPR MCP
tool is now using the short-lived MCP based tool which utilizes
Zendriver instead of Patchright.

There's a lot of work to be done for a generic Zendriver-based MCP tool,
but so far it's sufficient for NPR, etc. Note that these are the tools
used by Headline Hub.

To verify, launch the server as usual, connect MCP inspector to
/mcp-media, and invoke the tool to get the NPR headlines.