Zendriver: support authenticated HTTP proxy
Chromium/CDP only allows configuring the proxy server via the CLI. For
proxy username and password, we ought to install a special handler and
supply the credentials in the callback.

To verify, run the server with LOG_LEVEL=DEBUG and configured with some
sort of residential proxy. Then run Headline Hub (with Maxmind etc, to
get the IP geolocation) and use it to his this MCP GetGather. The log
should display that authenticated proxy is being set up for the traffic:

```
DEBUG    Supplying proxy authentication...
```