Implement distilled page using Zendriver (#681)
Only BBC tool (to get the saved article) is converted to this
Zendriver-based distill page.

To verify, launch MCP Inspector and connect to `/mcp-media` and use the
BBC tool `bbc_get_saved_articles`. Continue with the link given for the
distilled page to sign in with BBC email and password. After the sign-in
is completed, invoking the tool for the second time should display the
list of saved articles in JSON.

Because of the minor refactoring, it's also important to check that
there is no regression with the Patchright version of distilled page.
Hence, launch MCP Inspector to connect to `/mcp-books` and test
`goodreads_get_book_list` (it should still work as expected).